Move likelihoodModel module under lib/ and update imports
