+++ lib/likelihoodModel.py
